feature(backend) #173: split app.py into multiple files
