+++ backend/utils/__init__.py
@@ -1,0 +1,4 @@
+from .chunk_processor import ChunkProcessor
+from .file_handler import FileHandler
+
+__all__ = ['ChunkProcessor', 'FileHandler']
